requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
